requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
feedparser==6.0.11
requests==2.32.3
tenacity==8.5.0
python-dateutil==2.9.0.post0
